requirements.txt
certifi==2025.8.3 
charset_normalizer==3.4.3 
idna==3.10 
requests==2.32.5 
urllib3==2.5.0
python-dotenv==1.1.1
annotated-types==0.7.0 
anyio==4.10.0 
asyncio-throttle==1.0.2 
colorama==0.4.6 
distro==1.9.0 
h11==0.16.0 
httpcore==1.0.9 
httpx==0.28.1 
jiter==0.10.0 
numpy==2.3.2 
openai==1.102.0 
pandas==2.3.2 
pydantic==2.11.7 
pydantic-core==2.33.2 
python-dateutil==2.9.0.post0 
pytz==2025.2 
scipy==1.16.1 
six==1.17.0 
sniffio==1.3.1 
tqdm==4.67.1 
typing-extensions==4.15.0 
typing-inspection==0.4.1 
tzdata==2025.2
Mako==1.3.10 
MarkupSafe==3.0.2 
SQLAlchemy==2.0.43 
alembic==1.16.5 
click==8.2.1 
deprecation==2.1.0 
fastapi==0.116.1 
greenlet==3.2.4 
h2==4.3.0 
hpack==4.1.0 
hyperframe==6.1.0 
packaging==25.0 
postgrest==1.1.1 
psycopg2-binary==2.9.10 
pyjwt==2.10.1 
python-multipart==0.0.20 
realtime==2.7.0 
starlette==0.47.3 
storage3==0.12.1 
strenum==0.4.15 
supabase==2.18.1 
supabase-auth==2.12.3 
supabase-functions==0.10.1 
uvicorn==0.35.0 
websockets==15.0.1